requirements.txt
bcrypt==4.0.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dnspython
email-validator==2.0.0.post2
Flask
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
WTForms==3.1.0
zipp==3.17.0
